server/requirements.txt
Flask==1.1.2
google-cloud-datastore==1.15.1
google-cloud-logging==1.15.1
google-cloud-secret-manager==2.0.0
praw==7.1.0
pyfcm==1.4.7
